savepoint: modified savepoint test
